+++ apps/tempest.games/dev/https-dev.ts
@@ -1,4 +1,7 @@
 import { readFileSync } from "node:fs"
+import type { RequestListener, Server } from "node:http"
+import { createServer as createHttpServer } from "node:http"
+import { createServer as createSecureServer } from "node:https"
 import { resolve } from "node:path"
 
 import { env } from "../src/library/env"
@@ -11,3 +14,10 @@ export const httpsDev = env.VITE_USE_SELF_SIGNED_CERTIFICATE
 			key: readFileSync(resolve(devDir, `./key.pem`), `utf-8`),
 		}
 	: undefined
+
+export function createServer(listener: RequestListener): Server {
+	if (httpsDev) {
+		return createSecureServer(httpsDev, listener)
+	}
+	return createHttpServer({}, listener)
+}
