+++ apps/tempest.games/src/frontend/services/trpc-client-service.ts
@@ -1,6 +1,6 @@
 import { createTRPCClient, httpBatchLink } from "@trpc/client"
 
-import type { AppRouter } from "../../backend.bun"
+import type { AppRouter } from "../../backend/router"
 
 export const trpc = createTRPCClient<AppRouter>({
 	links: [
